# IndianMasterApp-Backend/docs/swagger.yaml
basePath: /api/v1
definitions:
  dto.APIResponse:
    properties:
      data: {}
      error: {}
      message:
        type: string
      meta:
        description: Pagination metadata or other metadata
      success:
        type: boolean
    type: object
...
  dto.JobSearchRequest:
    properties:
      benefits:
        type: string
      businessType:
        type: string
      city:
        type: string
      experienceMax:
        type: integer
      experienceMin:
        type: integer
      jobRole:
        type: string
      language:
        type: string
      salaryMax:
        type: number
      salaryMin:
        type: number
      workType:
        type: string
    type: object
...
  dto.WorkerSearchRequest:
    properties:
      availability:
        type: string
      city:
        type: string
      experienceMax:
        type: integer
      experienceMin:
        type: integer
      language:
        type: string
      role:
        type: string
      salaryMax:
        type: integer
      salaryMin:
        type: integer
      venuePreference:
        type: string
      workType:
        type: string
    type: object
...
info:
  contact:
    email: support@indianmaster.com
    name: API Support
  description: Backend APIs for Indian Master Application - A comprehensive platform
    for job seekers and hirers in India
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Indian Master API
  version: "1.0"
paths:
...
  /jobs/search:
    post:
      consumes:
      - application/json
      description: Perform advanced search on jobs with multiple filters
      parameters:
      - description: Search criteria
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.JobSearchRequest'
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Items per page (max 100)
        in: query
        name: limit
        type: integer
      - default: match_score
        description: Sort field (e.g., match_score, salary)
        in: query
        name: sort
        type: string
      - default: desc
        description: Sort order (asc or desc)
        enum:
        - asc
        - desc
        in: query
        name: order
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Jobs found successfully
          schema:
            $ref: '#/definitions/dto.APIResponse'
        "400":
          description: Invalid request parameters
          schema:
            $ref: '#/definitions/dto.APIResponse'
        "500":
          description: Failed to search jobs
          schema:
            $ref: '#/definitions/dto.APIResponse'
      summary: Search Jobs
      tags:
      - Search
  /notifications:
    get:
      consumes:
      - application/json
      description: Retrieve paginated notifications for the authenticated user (all,
        read + unread)
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Items per page (max 100)
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Notifications retrieved successfully
          schema:
            $ref: '#/definitions/dto.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.APIResponse'
        "500":
          description: Failed to fetch notifications
          schema:
            $ref: '#/definitions/dto.APIResponse'
      security:
      - BearerAuth: []
      summary: Get All Notifications
      tags:
      - Notification
  /notifications/{id}/read:
    patch:
      consumes:
      - application/json
      description: Mark a specific notification as read for the authenticated user
      parameters:
      - description: Notification ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Notification marked as read
          schema:
            $ref: '#/definitions/dto.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.APIResponse'
        "404":
          description: Notification not found
          schema:
            $ref: '#/definitions/dto.APIResponse'
        "500":
          description: Failed to mark notification as read
          schema:
            $ref: '#/definitions/dto.APIResponse'
      security:
      - BearerAuth: []
      summary: Mark Notification as Read
      tags:
      - Notification
  /notifications/unread:
    get:
      consumes:
      - application/json
      description: Retrieve paginated unread notifications for the authenticated user
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Items per page (max 100)
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Unread notifications retrieved successfully
          schema:
            $ref: '#/definitions/dto.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.APIResponse'
        "500":
          description: Failed to fetch unread notifications
          schema:
            $ref: '#/definitions/dto.APIResponse'
      security:
      - BearerAuth: []
      summary: Get Unread Notifications
      tags:
      - Notification
  /notifications/unread/count:
    get:
      consumes:
      - application/json
      description: Get the count of unread notifications for the authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: Unread count retrieved successfully
          schema:
            $ref: '#/definitions/dto.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.APIResponse'
        "500":
          description: Failed to get unread count
          schema:
            $ref: '#/definitions/dto.APIResponse'
      security:
      - BearerAuth: []
      summary: Get Unread Notification Count
      tags:
      - Notification
  /search/jobs:
    post:
      consumes:
      - application/json
      description: Perform advanced search on jobs with multiple filters. Results
        are cached in Redis for 5 minutes.
      parameters:
      - description: Search criteria
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.JobSearchRequest'
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Items per page (max 100)
        in: query
        name: limit
        type: integer
      - default: match_score
        description: Sort field (e.g., match_score, salary)
        in: query
        name: sort
        type: string
      - default: desc
        description: Sort order (asc or desc)
        enum:
        - asc
        - desc
        in: query
        name: order
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Jobs found successfully
          schema:
            $ref: '#/definitions/dto.APIResponse'
        "400":
          description: Invalid request parameters
          schema:
            $ref: '#/definitions/dto.APIResponse'
        "500":
          description: Failed to search jobs
          schema:
            $ref: '#/definitions/dto.APIResponse'
      security:
      - BearerAuth: []
      summary: Search Jobs (Cached)
      tags:
      - Search
  /search/workers:
    post:
      consumes:
      - application/json
      description: Perform advanced search on workers with multiple filters. Results
        are cached in Redis for 5 minutes.
      parameters:
      - description: Search criteria
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.WorkerSearchRequest'
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Items per page (max 100)
        in: query
        name: limit
        type: integer
      - default: match_score
        description: Sort field (e.g., match_score, experience_years)
        in: query
        name: sort
        type: string
      - default: desc
        description: Sort order (asc or desc)
        enum:
        - asc
        - desc
        in: query
        name: order
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Workers found successfully
          schema:
            $ref: '#/definitions/dto.APIResponse'
        "400":
          description: Invalid request parameters
          schema:
            $ref: '#/definitions/dto.APIResponse'
        "500":
          description: Failed to search workers
          schema:
            $ref: '#/definitions/dto.APIResponse'
      security:
      - BearerAuth: []
      summary: Search Workers (Cached)
      tags:
      - Search
...
  /workers/search:
    post:
      consumes:
      - application/json
      description: Perform advanced search on workers with multiple filters
      parameters:
      - description: Search criteria
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.WorkerSearchRequest'
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Items per page (max 100)
        in: query
        name: limit
        type: integer
      - default: match_score
        description: Sort field (e.g., match_score, experience_years)
        in: query
        name: sort
        type: string
      - default: desc
        description: Sort order (asc or desc)
        enum:
        - asc
        - desc
        in: query
        name: order
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Workers found successfully
          schema:
            $ref: '#/definitions/dto.APIResponse'
        "400":
          description: Invalid request parameters
          schema:
            $ref: '#/definitions/dto.APIResponse'
        "500":
          description: Failed to search workers
          schema:
            $ref: '#/definitions/dto.APIResponse'
      summary: Search Workers
      tags:
      - Search
schemes:
- http
- https
swagger: "2.0"